Basic support for folding when using 'foldmethod=syntax'
I'm vim user and from time to time I like to use folding (zc, zo
keystrokes).
The basic method I use is 'syntax', which isn't currently supported
by chpl. Note that setting foldmethod to 'indent' makes it work,
but it's inconvenient when using vim for multiple languages,
so the easiest solution is simply to add some lines for folding in syntax/chpl.vim.
Note that this could be expanded in future.